+++ react/tic-tac-toe/src/app/screens/Game/index.js
@@ -1,27 +1,10 @@
 import React, { Component } from 'react';
 
+import calculateWinner from '../../../utils/calculate-winner';
+
 import styles from './styles.module.scss';
 import Board from './components/Board';
 
-function calculateWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6]
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
-}
 
 class Game extends Component {
   state = {
